Broke game loop off into game manager
