+++ CS4450 Final Project/nbproject/private/private.xml
@@ -7,6 +7,8 @@
             <file>file:/C:/Users/Daniel/Documents/CPP/Fall2020/0A-CS4450/CS-4450-Final-Project/CS4450%20Final%20Project/src/Main.java</file>
             <file>file:/C:/Users/Daniel/Documents/CPP/Fall2020/0A-CS4450/CS-4450-Final-Project/CS4450%20Final%20Project/src/Chunk.java</file>
             <file>file:/C:/Users/Daniel/Documents/CPP/Fall2020/0A-CS4450/CS-4450-Final-Project/CS4450%20Final%20Project/src/Vector3Float.java</file>
+            <file>file:/C:/Users/Daniel/Documents/CPP/Fall2020/0A-CS4450/CS-4450-Final-Project/CS4450%20Final%20Project/terrain.png</file>
+            <file>file:/C:/Users/Daniel/Documents/CPP/Fall2020/0A-CS4450/CS-4450-Final-Project/CS4450%20Final%20Project/src/Block.java</file>
         </group>
     </open-files>
 </project-private>
