+++ CS4450 Final Project/src/FPCameraController.java
@@ -23,21 +23,6 @@ import org.lwjgl.Sys;
 
 public class FPCameraController {
     
-    private static final int ESCAPE = Keyboard.KEY_ESCAPE;
-
-    private static final int UP = Keyboard.KEY_SPACE;
-    private static final int DOWN = Keyboard.KEY_LSHIFT;
-    
-    private static final int FORWARD = Keyboard.KEY_W;
-    private static final int BACKWARD = Keyboard.KEY_S;
-    private static final int LEFT = Keyboard.KEY_A;
-    private static final int RIGHT = Keyboard.KEY_D;
-    
-    private static final int FORWARD_ARROW = Keyboard.KEY_UP;
-    private static final int BACKWARD_ARROW = Keyboard.KEY_DOWN;
-    private static final int LEFT_ARROW = Keyboard.KEY_LEFT;
-    private static final int RIGHT_ARROW = Keyboard.KEY_RIGHT;
-    
     // 3d vector to store the camera's position in
     private Vector3f position = null;
     private Vector3f lookPosition = null;
@@ -48,10 +33,6 @@ public class FPCameraController {
     // the rotation around the X axis of the camera
     private float pitch = 0.0f;
     
-    private Vector3Float me;
-
-	private Chunk chunk;
-
     /***************************************************************
     * method: FPCameraController
     * purpose: constructor used to instantiate the position Vector3f.
@@ -66,7 +47,6 @@ public class FPCameraController {
         lookPosition.y = 15f;
         lookPosition.z = 0f;
 
-	chunk = new Chunk((int)x, (int)y, (int)z);
     }
 
     /***************************************************************
@@ -199,158 +179,4 @@ public class FPCameraController {
         glLight(GL_LIGHT0, GL_POSITION, lightPosition);
     }
 
-    /***************************************************************
-    * method: gameLoop
-    * purpose: the primary game loop for this program.  Responsible for
-    *   calling relevant methods to respond to things like rendering, user
-    *   input, etc.
-    *
-    ****************************************************************/ 
-    public void gameLoop() {
-        FPCameraController camera = new FPCameraController(-30, -70, -30);
-        float dx = 0.0f;
-        float dy = 0.0f;
-        float dt = 0.0f; // length of frame
-        float lastTime = 0.0f; // when the last frame was
-        long time = 0;
-        float mouseSensitivity = 0.09f;
-        float movementSpeed = .35f;
-        // hide the mouse
-        Mouse.setGrabbed(true);
-
-		chunk.rebuildMesh();
-
-        // keep looping till the display window is closed the ESC key is down
-        while (!Display.isCloseRequested() && !Keyboard.isKeyDown(ESCAPE)) {
-            time = Sys.getTime();
-            lastTime = time;
-            
-            // distance in mouse movement
-            // from the last getDX() call.
-            dx = Mouse.getDX();
-            
-            // distance in mouse movement
-            // from the last getDY() call.
-            dy = Mouse.getDY();
-            
-            // control camera yaw from x movement front the mouse
-            camera.yaw(dx * mouseSensitivity);
-            
-            // control camera pitch from y movement front the mouse
-            camera.pitch(dy * mouseSensitivity);
-            
-            // when passing in the distance to move
-            // we times the movementSpeed with dt this is a time scale
-            // so if its a slow frame u move more then a fast frame
-            // so on a slow computer you move just as fast as on a fast computer
-            if (Keyboard.isKeyDown(FORWARD) || Keyboard.isKeyDown(FORWARD_ARROW))// move forward
-            {
-                camera.walkForward(movementSpeed);
-            }
-            
-            if (Keyboard.isKeyDown(BACKWARD) || Keyboard.isKeyDown(BACKWARD_ARROW))// move backwards
-            {
-                camera.walkBackwards(movementSpeed);
-            }
-            
-            if (Keyboard.isKeyDown(LEFT) || Keyboard.isKeyDown(LEFT_ARROW)) {// strafe left {
-                camera.strafeLeft(movementSpeed);
-            }
-            
-            if (Keyboard.isKeyDown(RIGHT) || Keyboard.isKeyDown(RIGHT_ARROW))// strafe right {
-            {
-                camera.strafeRight(movementSpeed);
-            }
-            
-            if (Keyboard.isKeyDown(UP)) // move up
-            {
-                camera.moveUp(movementSpeed);
-            }
-
-            if (Keyboard.isKeyDown(DOWN)) {
-                camera.moveDown(movementSpeed);
-            }
-            
-            // set the modelview matrix back to the identity
-            glLoadIdentity();
-            
-            // look through the camera before you draw anything
-            camera.lookThrough();
-            glClear(GL_COLOR_BUFFER_BIT | GL_DEPTH_BUFFER_BIT);
-            
-            // you would draw your scene here.
-			chunk.render();
-            //render();
-            
-            // draw the buffer to the screen
-            Display.update();
-            Display.sync(60);
-        }
-        Display.destroy();
-    }
-
-    /***************************************************************
-    * method: render
-    * purpose: renders shapes into the 3D space.
-    *
-    ****************************************************************/ 
-    private void render() {
-        try {
-            glBegin(GL_QUADS);
-            glColor3f(1.0f, 0.0f, 1.0f);
-            
-            glVertex3f(1.0f, -1.0f, -1.0f);
-            glVertex3f(-1.0f, -1.0f, -1.0f);
-            glVertex3f(-1.0f, 1.0f, -1.0f);
-            glVertex3f(1.0f, 1.0f, -1.0f);
-            glEnd();
-            
-            glBegin(GL_QUADS);
-            glColor3f(0.7f, 0.2f, 1.0f);
-            
-            glVertex3f(1.0f, -1.0f, -3.0f);
-            glVertex3f(-1.0f, -1.0f, -3.0f);
-            glVertex3f(-1.0f, 1.0f, -3.0f);
-            glVertex3f(1.0f, 1.0f, -3.0f);
-            glEnd();
-            
-            glBegin(GL_QUADS);
-            glColor3f(0.0f, 0.0f, 1.0f);
-            
-            glVertex3f(1.0f, -1.0f, -1.0f);
-            glVertex3f(1.0f, -1.0f, -3.0f);
-            glVertex3f(1.0f, 1.0f, -3.0f);
-            glVertex3f(1.0f, 1.0f, -1.0f);
-            glEnd();
-            
-            glBegin(GL_QUADS);
-            glColor3f(1.0f, 0.0f, 0.0f);
-            
-            glVertex3f(-1.0f, -1.0f, -3.0f);
-            glVertex3f(-1.0f, -1.0f, -1.0f);
-            glVertex3f(-1.0f, 1.0f, -1.0f);
-            glVertex3f(-1.0f, 1.0f, -3.0f);
-            glEnd();
-            
-            glBegin(GL_QUADS);
-            glColor3f(1.0f, 0.5f, 0.5f);
-            
-            glVertex3f(1.0f, -1.0f, -1.0f);
-            glVertex3f(-1.0f, -1.0f, -1.0f);
-            glVertex3f(-1.0f, -1.0f, -3.0f);
-            glVertex3f(1.0f, -1.0f, -3.0f);
-            glEnd();
-            
-            glBegin(GL_QUADS);
-            glColor3f(0.0f, 1.0f, 0.0f);
-            
-            glVertex3f(1.0f, 1.0f, -1.0f);
-            glVertex3f(-1.0f, 1.0f, -1.0f);
-            glVertex3f(-1.0f, 1.0f, -3.0f);
-            glVertex3f(1.0f, 1.0f, -3.0f);
-            glEnd();
-
-            
-        } catch (Exception e) { }
-    }
 }
